refactor: move the common ConditionalEqGadget impls above int and uint
