+++ gadgets/src/traits/utilities/int/utilities/mod.rs
@@ -18,8 +18,5 @@
 pub mod alloc;
 pub use self::alloc::*;
 
-pub mod eq;
-pub use self::eq::*;
-
 pub mod select;
 pub use self::select::*;
